Remove src/plugins/visualizations -> src/plugins/visualize cyclic dependencies
Part of #84750
