+++ src/plugins/visualizations/public/index.ts
@@ -44,6 +44,9 @@ export type {
   ReactVisTypeOptions,
   Schema,
   ISchemas,
+  VisEditorConstructor,
+  IEditorController,
+  EditorRenderProps,
 } from './vis_types';
 export { VisParams, SerializedVis, SerializedVisData, VisData } from './vis';
 export type VisualizeEmbeddableFactoryContract = PublicContract<VisualizeEmbeddableFactory>;
