+++ src/plugins/visualizations/public/vis_types/index.ts
@@ -20,6 +20,13 @@
 export * from './types_service';
 export { Schemas } from './schemas';
 export { VisGroups } from './types';
-export type { VisType, ISchemas, Schema } from './types';
+export type {
+  VisType,
+  ISchemas,
+  Schema,
+  IEditorController,
+  VisEditorConstructor,
+  EditorRenderProps,
+} from './types';
 export type { BaseVisTypeOptions } from './base_vis_type';
 export type { ReactVisTypeOptions } from './react_vis_type';
