+++ src/plugins/visualizations/public/vis_types/types.ts
@@ -16,13 +16,23 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-
+import { EventEmitter } from 'events';
 import { IconType } from '@elastic/eui';
 import React, { ReactNode } from 'react';
 import { Adapters } from 'src/plugins/inspector';
-import { VisEditorConstructor } from 'src/plugins/visualize/public';
-import { IndexPattern, AggGroupNames, AggParam, AggGroupName } from '../../../data/public';
+import { CoreStart, SavedObject } from 'src/core/public';
+import {
+  IndexPattern,
+  AggGroupNames,
+  AggParam,
+  AggGroupName,
+  DataPublicPluginStart,
+  Filter,
+  TimeRange,
+  Query,
+} from '../../../data/public';
 import { Vis, VisParams, VisToExpressionAst, VisualizationControllerConstructor } from '../types';
+import { PersistedState, VisualizeEmbeddableContract } from '../index';
 
 export interface VisTypeOptions {
   showTimePicker: boolean;
@@ -151,4 +161,30 @@ export interface VisType<TVisParams = unknown> {
   // TODO: The following types still need to be refined properly.
   readonly editorConfig: Record<string, any>;
   readonly visConfig: Record<string, any>;
+}
+
+export type VisEditorConstructor = new (
+  element: HTMLElement,
+  vis: Vis,
+  eventEmitter: EventEmitter,
+  embeddableHandler: VisualizeEmbeddableContract
+) => IEditorController;
+
+export interface IEditorController {
+  render(props: EditorRenderProps): Promise<void> | void;
+  destroy(): void;
+}
+
+export interface EditorRenderProps {
+  core: CoreStart;
+  data: DataPublicPluginStart;
+  filters: Filter[];
+  timeRange: TimeRange;
+  query?: Query;
+  savedSearch?: SavedObject;
+  uiState: PersistedState;
+  /**
+   * Flag to determine if visualiztion is linked to the saved search
+   */
+  linked: boolean;
 }
