+++ src/plugins/visualize/public/application/types.ts
@@ -18,9 +18,8 @@
  */
 
 import { History } from 'history';
-import { TimeRange, Query, Filter, DataPublicPluginStart } from 'src/plugins/data/public';
+import { Query, Filter, DataPublicPluginStart } from 'src/plugins/data/public';
 import {
-  PersistedState,
   SavedVisState,
   VisualizationsStart,
   Vis,
@@ -45,7 +44,6 @@ import { SharePluginStart } from 'src/plugins/share/public';
 import { SavedObjectsStart, SavedObject } from 'src/plugins/saved_objects/public';
 import { EmbeddableStart, EmbeddableStateTransfer } from 'src/plugins/embeddable/public';
 import { UrlForwardingStart } from 'src/plugins/url_forwarding/public';
-import { EventEmitter } from 'events';
 import { DashboardStart } from '../../../dashboard/public';
 import type { SavedObjectsTaggingApi } from '../../../saved_objects_tagging_oss/public';
 
@@ -79,20 +77,6 @@ export type VisualizeAppStateContainer = ReduxLikeStateContainer<
   VisualizeAppState,
   VisualizeAppStateTransitions
 >;
-
-export interface EditorRenderProps {
-  core: CoreStart;
-  data: DataPublicPluginStart;
-  filters: Filter[];
-  timeRange: TimeRange;
-  query?: Query;
-  savedSearch?: SavedObject;
-  uiState: PersistedState;
-  /**
-   * Flag to determine if visualiztion is linked to the saved search
-   */
-  linked: boolean;
-}
 
 export interface VisualizeServices extends CoreStart {
   stateTransferService: EmbeddableStateTransfer;
@@ -135,15 +119,3 @@ export interface ByValueVisInstance {
 }
 
 export type VisualizeEditorVisInstance = SavedVisInstance | ByValueVisInstance;
-
-export type VisEditorConstructor = new (
-  element: HTMLElement,
-  vis: Vis,
-  eventEmitter: EventEmitter,
-  embeddableHandler: VisualizeEmbeddableContract
-) => IEditorController;
-
-export interface IEditorController {
-  render(props: EditorRenderProps): Promise<void> | void;
-  destroy(): void;
-}
